# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sysdesign API",
    "description": "Thin API over the Module 1 influencer-intelligence schema. FastAPI generates the OpenAPI spec automatically: machine-readable at /openapi.json, Swagger UI at /docs, ReDoc at /redoc. The Module 2 frontend codegens a typed client from that spec.",
    "version": "0.2.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health"
                }
              }
            }
          }
        }
      }
    },
    "/influencers": {
      "get": {
        "tags": [
          "influencers"
        ],
        "summary": "List Influencers",
        "operationId": "list_influencers",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Influencer"
                  },
                  "type": "array",
                  "title": "Response List Influencers"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "influencers"
        ],
        "summary": "Create Influencer",
        "operationId": "create_influencer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InfluencerIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Influencer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/influencers/{influencer_id}": {
      "patch": {
        "tags": [
          "influencers"
        ],
        "summary": "Update Watermark",
        "description": "Advance the incremental-scrape watermark. The scraper calls this after a run so the\nnext run only pulls posts newer than last_scraped_at.",
        "operationId": "update_watermark",
        "parameters": [
          {
            "name": "influencer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Influencer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InfluencerWatermark"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Influencer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List Sources",
        "operationId": "list_sources",
        "parameters": [
          {
            "name": "influencer_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Influencer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Source"
                  },
                  "title": "Response List Sources"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Create Source",
        "operationId": "create_source",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Source"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/signals": {
      "post": {
        "tags": [
          "signals"
        ],
        "summary": "Create Signal",
        "description": "Idempotent write. content_hash is derived here, so identical payloads dedupe no\nmatter who sends them. Returns inserted=False when the exact signal already existed.",
        "operationId": "create_signal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignalIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalInsertResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "List Signals",
        "description": "A time window is REQUIRED so the query always carries the partition key and prunes.\nFiltering by influencer alone (no window) would fan out across every partition, so the\nAPI simply doesn't offer that shape.",
        "operationId": "list_signals",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "start of window (inclusive)",
              "title": "From"
            },
            "description": "start of window (inclusive)"
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "end of window (exclusive)",
              "title": "To"
            },
            "description": "end of window (exclusive)"
          },
          {
            "name": "influencer_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Influencer Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Signal"
                  },
                  "title": "Response List Signals"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rollup": {
      "get": {
        "tags": [
          "rollup"
        ],
        "summary": "Rollup",
        "description": "Reads the precomputed daily_signal_rollup matview, not raw_signals. The dashboard\nnever pays for the aggregate. Refresh happens out of band (make seed / a scheduled job).",
        "operationId": "rollup",
        "parameters": [
          {
            "name": "influencer_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Influencer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyRollup"
                  },
                  "title": "Response Rollup"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DailyRollup": {
        "properties": {
          "influencer_id": {
            "type": "integer",
            "title": "Influencer Id"
          },
          "day": {
            "type": "string",
            "format": "date-time",
            "title": "Day"
          },
          "signal_count": {
            "type": "integer",
            "title": "Signal Count"
          },
          "source_count": {
            "type": "integer",
            "title": "Source Count"
          }
        },
        "type": "object",
        "required": [
          "influencer_id",
          "day",
          "signal_count",
          "source_count"
        ],
        "title": "DailyRollup"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Influencer": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "instagram_handle": {
            "type": "string",
            "title": "Instagram Handle"
          },
          "last_scraped_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Scraped At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "instagram_handle",
          "last_scraped_at",
          "created_at"
        ],
        "title": "Influencer"
      },
      "InfluencerIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "instagram_handle": {
            "type": "string",
            "title": "Instagram Handle"
          }
        },
        "type": "object",
        "required": [
          "name",
          "instagram_handle"
        ],
        "title": "InfluencerIn"
      },
      "InfluencerWatermark": {
        "properties": {
          "last_scraped_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Scraped At"
          }
        },
        "type": "object",
        "required": [
          "last_scraped_at"
        ],
        "title": "InfluencerWatermark"
      },
      "Signal": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "influencer_id": {
            "type": "integer",
            "title": "Influencer Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "id",
          "influencer_id",
          "source_id",
          "captured_at",
          "content_hash",
          "payload"
        ],
        "title": "Signal"
      },
      "SignalIn": {
        "properties": {
          "influencer_id": {
            "type": "integer",
            "title": "Influencer Id"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          }
        },
        "type": "object",
        "required": [
          "influencer_id",
          "captured_at",
          "payload"
        ],
        "title": "SignalIn"
      },
      "SignalInsertResult": {
        "properties": {
          "inserted": {
            "type": "boolean",
            "title": "Inserted"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          }
        },
        "type": "object",
        "required": [
          "inserted",
          "content_hash"
        ],
        "title": "SignalInsertResult"
      },
      "Source": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "influencer_id": {
            "type": "integer",
            "title": "Influencer Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "influencer_id",
          "kind",
          "url",
          "created_at"
        ],
        "title": "Source"
      },
      "SourceIn": {
        "properties": {
          "influencer_id": {
            "type": "integer",
            "title": "Influencer Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "influencer_id",
          "kind",
          "url"
        ],
        "title": "SourceIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "health",
      "description": "Liveness + db reachability."
    },
    {
      "name": "influencers",
      "description": "The creators we track (Defrag watchlist)."
    },
    {
      "name": "sources",
      "description": "Where an influencer's signals come from."
    },
    {
      "name": "signals",
      "description": "The partitioned raw_signals firehose."
    },
    {
      "name": "rollup",
      "description": "Precomputed daily counts (materialized view)."
    }
  ]
}